+++ lindera/src/token_filter.rs
@@ -1,4 +1,5 @@
 pub mod japanese_base_form;
+pub mod japanese_compound_word;
 pub mod japanese_katakana_stem;
 pub mod japanese_keep_tags;
 pub mod japanese_reading_form;