+++ lindera/src/token_filter/keep_words.rs
@@ -9,12 +9,12 @@ use crate::{error::LinderaErrorKind, LinderaResult, Token};
 pub const KEEP_WORDS_TOKEN_FILTER_NAME: &str = "keep_words";
 #[derive(Clone, Debug, Deserialize, Serialize, PartialEq, Eq)]
 pub struct KeepWordsTokenFilterConfig {
-    keep_words: HashSet<String>,
+    words: HashSet<String>,
 }
 
 impl KeepWordsTokenFilterConfig {
-    pub fn new(keep_words: HashSet<String>) -> Self {
-        Self { keep_words }
+    pub fn new(words: HashSet<String>) -> Self {
+        Self { words }
     }
 
     pub fn from_slice(data: &[u8]) -> LinderaResult<Self> {
@@ -43,7 +43,8 @@ impl TokenFilter for KeepWordsTokenFilter {
     }
 
     fn apply<'a>(&self, tokens: &mut Vec<Token<'a>>) -> LinderaResult<()> {
-        tokens.retain(|token| self.config.keep_words.contains(token.text.as_ref()));
+        tokens.retain(|token| self.config.words.contains(token.text.as_ref()));
+
         Ok(())
     }
 }
@@ -63,7 +64,7 @@ mod tests {
     fn test_keep_words_token_filter_config_from_slice() {
         let config_str = r#"
         {
-            "keep_words": [
+            "words": [
                 "Lindera",
                 "Rust"
             ]
@@ -71,14 +72,14 @@ mod tests {
         "#;
         let config = KeepWordsTokenFilterConfig::from_slice(config_str.as_bytes()).unwrap();
 
-        assert_eq!(config.keep_words.len(), 2);
+        assert_eq!(config.words.len(), 2);
     }
 
     #[test]
     fn test_keep_words_token_filter_from_slice() {
         let config_str = r#"
         {
-            "keep_words": [
+            "words": [
                 "Lindera",
                 "Rust"
             ]
@@ -93,7 +94,7 @@ mod tests {
     fn test_keep_words_token_filter_apply() {
         let config_str = r#"
         {
-            "keep_words": [
+            "words": [
                 "Lindera",
                 "Rust"
             ]